requirements.txt
contourpy==1.3.3
cycler==0.12.1
dnspython==2.8.0
fonttools==4.61.1
greenlet==3.2.4
kiwisolver==1.4.9
matplotlib==3.10.8
mplfinance==0.12.10b0
narwhals==2.13.0
numpy==2.3.5
packaging==25.0
pandas==2.3.3
pillow==12.0.0
playwright==1.55.0
plotly==6.5.0
pyee==13.0.0
pymongo==4.15.5
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
typing_extensions==4.15.0
tzdata==2025.3
questdb==4.1.0
psycopg2-binary==2.9.9
requests==2.31.0
Flask==2.2.2
pyzmq
upstox-python-sdk

fastapi
uvicorn
